% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DiffModification}
\alias{DiffModification}
\title{Differential modification by linear model test}
\usage{
DiffModification(
  object,
  col.data,
  contrast.df,
  norm.method = NULL,
  test.gene.id = NULL,
  assay.colname = "assay",
  biology.colname = "bio",
  input.id = "Input",
  enrich.id = "Enrich",
  norm = FALSE,
  adjust = TRUE,
  p.cutoff = 0.05,
  pseudo.count = 0.125,
  eps = 1e-04,
  shrink_lambda = 0
)
}
\arguments{
\item{object}{}

\item{col.data}{\code{data.frame} with at least three columns (indicate condition, enrich and biological groups).
Rows of \code{col.data} correspond to columns of \code{data}.}

\item{contrast.df}{Data frame of contrast, where extracting results as
first column vs. second column.}

\item{norm.method}{Normalization methods to use}

\item{test.gene.id}{List of gene ids to be tested}

\item{assay.colname}{Name of the assay column in \code{col.data}, e.g., "enrich".}

\item{biology.colname}{Name of the biological column in \code{col.data}, e.g., "bio".}

\item{input.id}{Input library id, must be consistent with the enrich column of \code{col.data}, e.g., "Input".}

\item{enrich.id}{Enrich library id, must be consistent with the enrich column of \code{col.data}, e.g., "Enrich".}

\item{adjust}{Whether to perform adjustment based on input expression levels.}

\item{p.cutoff}{P values Cutoff for significant differential modification genes.}

\item{pseudo.count}{Minimal offset to prevent negative values.}

\item{eps}{Threshold to prevent negative values, default: 1e-4}

\item{shrink_lambda}{Factor to control the shrinkage of expression size factor from input, default: 0.}
}
\value{
List of differential modification results
}
\description{
Differential modification by linear model test
}
