% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.expressionFactor}
\alias{.expressionFactor}
\title{Size factor based on input expression}
\usage{
.expressionFactor(data_input, eps = 1e-04)
}
\arguments{
\item{data_input}{}

\item{eps}{see description in \code{DiffModification}}
}
\value{
vector of expression factors
}
\description{
Size factor based on input expression
}
