% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.mergeDiffModWithEnrich}
\alias{.mergeDiffModWithEnrich}
\title{Merge differential modification data with enrichment data}
\usage{
.mergeDiffModWithEnrich(
  diffmod_res,
  enrich_list,
  contrast_name,
  fc.thresh = log2(1.2),
  p.thresh = 0.05
)
}
\arguments{
\item{diffmod_res}{Table of differential modification results from \code{DiffModification}.}

\item{enrich_list}{Enriched genes from \code{FindEnrichment}}

\item{contrast_name}{Contrast name, e.g., "A_B".}

\item{fc.thresh}{Fold-change threshold for significant differential modification.}

\item{p.thresh}{P values threshold for significant differential modification.}
}
\value{
list
}
\description{
Merge differential modification data with enrichment data
}
