% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.rmzero}
\alias{.rmzero}
\title{Remove 0 from vector}
\usage{
.rmzero(x, eps = 1e-04)
}
\arguments{
\item{x}{}

\item{eps}{see description in \code{DiffModification}}
}
\value{
vector
}
\description{
Remove 0 from vector
}
