% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.safe_grep}
\alias{.safe_grep}
\title{Safe grep}
\usage{
.safe_grep(pattern, x, ...)
}
\arguments{
\item{pattern}{Same as \code{grep}. character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector.}

\item{x}{Same as \code{grep}. a character vector where matches are sought, or an object which can be coerced by as.character to a character vector.}

\item{...}{Other parameters passed to \code{grep}}
}
\value{
matched value
}
\description{
Safe grep
}
