% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.shrinkClip}
\alias{.shrinkClip}
\title{Shrink factors toward 1 and clip bounds}
\usage{
.shrinkClip(
  size_factor,
  shrink_lambda = 0,
  min_factor = 1e-06,
  max_factor = 1e+06
)
}
\arguments{
\item{size_factor}{from \code{.expressionFactor}.}

\item{shrink_lambda}{see description in \code{DiffModification}.}

\item{min_factor}{minimum factor to clip, default: 1e-6.}

\item{max_factor}{maximum factor to clip, default: 1e6.}
}
\value{
shrunk expression factors
}
\description{
Shrink factors toward 1 and clip bounds
}
