% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getFactor}
\alias{getFactor}
\alias{getFactor,Gretool,character,character-method}
\title{Accessor of normalization factors}
\usage{
getFactor(object, slot = c("sample", "spike_in"), method)

\S4method{getFactor}{Gretool,character,character}(object, slot = c("sample", "spike_in"), method)
}
\arguments{
\item{object}{Gretool.}

\item{slot}{Which slot to get, one of \code{sample} or \code{spike_in}.}

\item{method}{Which normalization methods to get, must be one of the methods presented in the selected slot.}
}
\value{
vector or list of factors.
}
\description{
Accessor of normalization factors
}
