% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getGeneSet}
\alias{getGeneSet}
\alias{getGeneSet,Gretool,character-method}
\title{Accessor of gene set}
\usage{
getGeneSet(object, name = c("NegControl", "NegEvaluation", "PosEvaluation"))

\S4method{getGeneSet}{Gretool,character}(object, name = c("NegControl", "NegEvaluation", "PosEvaluation"))
}
\arguments{
\item{object}{Gretool.}

\item{name}{Name of the gene set, one of \code{NegControl},
\code{NegEvaluation}, or \code{PosEvaluation}}
}
\value{
Vector of gene ID
}
\description{
Get gene set.
}
