% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getMetrics}
\alias{getMetrics}
\alias{getMetrics,Gretool-method}
\title{Accessor of metrics}
\usage{
getMetrics(object)

\S4method{getMetrics}{Gretool}(object)
}
\arguments{
\item{object}{Gretool.}
}
\value{
data.frame
}
\description{
Accessor of metrics
}
