% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{listParameter}
\alias{listParameter}
\alias{getParameter}
\alias{listParameter,Gretool-method}
\alias{getParameter,Gretool,character-method}
\title{Accessor of parameters}
\usage{
listParameter(object)

getParameter(object, name)

\S4method{listParameter}{Gretool}(object)

\S4method{getParameter}{Gretool,character}(object, name)
}
\arguments{
\item{object}{Gretool.}

\item{name}{Name of the parameter.}
}
\value{
Vector of parameter names

Vector of parameter
}
\description{
List all parameters.

Get specific parameters.
}
