% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getScore}
\alias{getScore}
\alias{getScore,Gretool-method}
\title{Accessor of normalization score}
\usage{
getScore(object)

\S4method{getScore}{Gretool}(object)
}
\arguments{
\item{object}{Gretool.}
}
\value{
data.frame
}
\description{
Accessor of normalization score
}
