% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{listNormalization}
\alias{listNormalization}
\alias{listNormalization,Gretool-method}
\title{Accessor of normalization methods}
\usage{
listNormalization(object)

\S4method{listNormalization}{Gretool}(object)
}
\arguments{
\item{object}{Gretool.}
}
\value{
Vector of normalization methods
}
\description{
List all normalization
}
